Multicore setup, fix laggy display render, setup for player response

+++ include/lights/led-buffer.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 #ifdef RELEASE
 #define FASTLED_INTERNAL
 #include <FastLED.h>
@@ -14,14 +16,14 @@ namespace Lights
   class LedBuffer
   {
   public:
-    LedBuffer(unsigned int numLeds) : _size{numLeds}, leds{new CRGB[numLeds]} {}
+    LedBuffer(uint16_t numLeds) : _size{numLeds}, leds{new CRGB[numLeds]} {}
     LedBuffer() { delete leds; }
     LedBuffer(LedBuffer &&other) = delete;
     LedBuffer &operator=(LedBuffer &&other) = delete;
 
     LedBuffer(const LedBuffer &other) : _size(other._size), leds(new CRGB[other._size])
     {
-      for (unsigned int i = 0; i < _size; ++i)
+      for (uint16_t i = 0; i < _size; ++i)
       {
         leds[i] = other.leds[i];
       }
@@ -32,7 +34,7 @@ namespace Lights
       if (this != &other)
       {
         CRGB *newLeds = new CRGB[other._size];
-        for (unsigned int i = 0; i < other._size; ++i)
+        for (uint16_t i = 0; i < other._size; ++i)
         {
           newLeds[i] = other.leds[i];
         }
@@ -44,16 +46,16 @@ namespace Lights
       return *this;
     }
 
-    CRGB &operator[](unsigned int index) { return leds[index]; }
-    const CRGB &operator[](unsigned int index) const { return leds[index]; }
+    CRGB &operator[](uint16_t index) { return leds[index]; }
+    const CRGB &operator[](uint16_t index) const { return leds[index]; }
 
     explicit operator CRGB *() { return leds; }
     explicit operator const CRGB *() const { return leds; }
 
-    unsigned int size() { return _size; }
+    uint16_t size() { return _size; }
 
   private:
-    unsigned int _size;
+    uint16_t _size;
     CRGB *leds;
   };
 }
